Apply fixes from ISSUES-AND-FIXES.md
- #8: Map YAML 'int' to C# 'long' (DOM only supports Int64)
- #9: Unwrap array-of-string ApiTriggerInput in generated UDAPI RunSafe
- #4: Unwrap array-wrapped parameter values in Aspire ApiService mock
- #10: Prevent route prefix doubling when apiRoute already ends with model name
- #6+7: Update frontend tester to use SAML cookie auth and correct CSS selectors

+++ dataminer-devpack-builder/New-DevPack.cs
@@ -809,7 +809,7 @@ static string GetCSharpType(PropertyConfig prop)
     {
         "string" => "string",
         "DateTime" => "DateTime",
-        "int" => "int",
+        "int" => "long",  // DataMiner DOM only supports Int64, not Int32
         "double" => "double",
         "bool" => "bool",
         _ => prop.Type
